Move appConfig within src and switch to module imports
