+++ src/core/Analytics.ts
@@ -1,9 +1,8 @@
 import * as Amplitude from 'expo-analytics-amplitude';
 import Constants from 'expo-constants';
 
-import appConfig from '../../appConfig';
-
-import UserService from './user/UserService';
+import appConfig from '@covid/appConfig';
+import UserService from '@covid/core/user/UserService';
 
 let isInitialized = false;
